+++ packages/react/src/components/FieldSet/index.ts
@@ -1,2 +1,2 @@
-export { FieldSet, FieldSetSize } from './FieldSet';
+export { FieldSet } from './FieldSet';
 export type { FieldSetProps } from './FieldSet';
