+++ packages/react/src/components/RadioButton/RadioButton.tsx
@@ -6,11 +6,6 @@ import { CheckboxRadioTemplate } from '../_CheckboxRadioTemplate';
 
 import classes from './RadioButton.module.css';
 
-export enum RadioButtonSize {
-  Xsmall = 'xsmall',
-  Small = 'small',
-}
-
 export interface RadioButtonProps {
   checked?: boolean;
   className?: string;
@@ -24,7 +19,7 @@ export interface RadioButtonProps {
   onChange: ChangeEventHandler<HTMLInputElement>;
   presentation?: boolean;
   radioId?: string;
-  size?: RadioButtonSize;
+  size?: 'small' | 'xsmall';
   value: string;
 }
 
@@ -40,7 +35,7 @@ const RadioButton = ({
   onChange,
   presentation,
   radioId,
-  size = RadioButtonSize.Small,
+  size = 'small',
   value,
 }: RadioButtonProps) => (
   <CheckboxRadioTemplate
@@ -61,7 +56,7 @@ const RadioButton = ({
     name={name}
     onChange={onChange}
     presentation={presentation}
-    size={size === RadioButtonSize.Xsmall ? 'xsmall' : 'small'}
+    size={size}
     type='radio'
     value={value}
   >
