+++ packages/react/src/components/index.ts
@@ -4,18 +4,18 @@ export { ErrorMessage } from './ErrorMessage';
 export type { ErrorMessageProps } from './ErrorMessage';
 export { Tabs } from './Tabs';
 export type { TabsProps, TabItem } from './Tabs';
-export { FieldSet, FieldSetSize } from './FieldSet';
+export { FieldSet } from './FieldSet';
 export type { FieldSetProps } from './FieldSet';
 export { Checkbox } from './Checkbox';
 export type { CheckboxProps } from './Checkbox';
 export { CheckboxGroup } from './CheckboxGroup';
 export type { CheckboxGroupProps, CheckboxGroupItem } from './CheckboxGroup';
-export { HelpText, HelpTextSize } from './HelpText';
+export { HelpText } from './HelpText';
 export { Popover, PopoverVariant } from './Popover';
 export type { PopoverProps } from './Popover';
-export { RadioButton, RadioButtonSize } from './RadioButton';
+export { RadioButton } from './RadioButton';
 export type { RadioButtonProps } from './RadioButton';
-export { RadioGroup, RadioGroupSize, RadioGroupVariant } from './RadioGroup';
+export { RadioGroup, RadioGroupVariant } from './RadioGroup';
 export type { RadioGroupProps, RadioGroupItem } from './RadioGroup';
 export { Select } from './Select';
 export type {
